get_menu + gitignore

+++ menu/get_Menu.py
@@ -3,6 +3,8 @@ from json import dump, load
 from datetime import datetime
 from re import search, escape
 from pdf2image import convert_from_bytes
+from os.path import isfile
+from os import remove, stat
 
 date = str(datetime.now().date())
 date = date[-2:] + '.' + date[-5:-3] + '.' + date[-8:-6]
@@ -22,16 +24,26 @@ def downloadMenu():
     return names
 
 def lastDownloadDate():
-    with open('data.json', 'r') as file:
-        if file.read():
+    if stat('data.json').st_size != 0:
+        with open('data.json', 'r') as file:
             return load(file)['date']
     return None
 
+def lastNames():
+    if stat('data.json').st_size != 0:
+        with open('data.json', 'r') as file:
+            return load(file)['names']
+    return []
+
 def getMenu():
-    with open('data.json', 'w') as file:
-        pass
+    if not isfile('data.json'):
+        with open('data.json', 'w') as file:
+            pass
     if lastDownloadDate() != date:
+        for i in lastNames():
+            remove(i)
         with open('data.json', 'w') as file:
             names = downloadMenu()
             data = {'date': date, 'names': names}
             dump(data, file)
+    return lastNames()
